requirements.txt
Flask==2.3.2
opencv-python==4.7.0.72
numpy==1.21.6
torch==1.13.1
torchvision==0.14.1
pillow==9.2.0
werkzeug==2.3.2
